+++ app/screens/createCategory/index.tsx
@@ -3,11 +3,9 @@ import { Input, ScrollView } from "native-base";
 import { Text, View } from "react-native";
 import { NativeStackScreenProps } from "@react-navigation/native-stack"
 
-import ImageBase64 from "../../models/interfaces/ImageBase64";
 import RootStackParamList from "../../models/interfaces/RootScreensParams"
 
 import { BottomButton, ImageSelector } from "../../components";
-import { listImgBase64 } from "../../resources/static/categoriesImages";
 
 import { useStyles } from "./styles";
 import useCreateCategory from "./hooks/useCreateCategory";
@@ -30,22 +28,21 @@ const CreateCategoryScreen: React.FC<CreateCategoryScreenProps> = ({ route, navi
           marginX={5}
           value={category?.name}
           variant="underlined" 
-          placeholder="Name" 
+          placeholder="Name*" 
           onChangeText={(text: string) => onChange('name', text)}
         />
         <Input 
           fontSize={20} 
           marginTop={10}
-          marginBottom={30}
           marginX={5}
           value={category?.budget?.toString()}
           variant="underlined" 
-          placeholder="Budget" 
+          placeholder="Budget*" 
           keyboardType='numeric'
           returnKeyType='done'
           onChangeText={(text: string) => onChange('budget', text)} 
         />
-        <ImageSelector listImageInfo={listImgBase64} selectedImage={listImgBase64.find((item) => item.id === paramCategory?.iconId)} onPress={(item: ImageBase64) => onChange('iconId', item.id)} />
+        <ImageSelector selectedIcon={paramCategory?.iconName || paramCategory?.iconId} onPress={(item: string) => onChange('iconName', item)} />
       </ScrollView>
       <BottomButton title={paramCategory ? 'Update' : 'Create'} onButtonPress={onButtonPress} />
     </View>